extract payment info into its own component
